+++ tools/register_applets.cxx
@@ -29,7 +29,8 @@ namespace tools {
 #ifdef VIAME_TOOLS_ENABLE_PYTHON
 
 VIAME_PYTHON_SCRIPT_APPLET( process_video_applet, "run",
-  "process_video.py", "Process new videos" )
+  "process_video.py",
+  "Process videos or images, or run a single pipeline file." )
 
 VIAME_PYTHON_SCRIPT_APPLET( add_segmentations_applet, "add-segmentations",
   "add_segmentations.py",